Fixed Many things
Player movement is fixed even if 100 scenes are loaded.
Game Over appears only when he drops in the water, but not when he loses
from Enemy.
Pause screen still appears randomly, but rarerly, but if i can fix it by
tonight.
Game over screen has a replay button instead of the ones in pause or
main screen.
Score is working and is shown on top of screen and in game over.
Score continues to work when you get to next level and resets if you
loose. Same with health.